+++ test/unit/gen_libs/code_coverage.sh
@@ -10,6 +10,7 @@ echo ""
 echo "Running unit test modules in conjunction with coverage"
 coverage run -a --source=gen_libs test/unit/gen_libs/add_cmd.py
 coverage run -a --source=gen_libs test/unit/gen_libs/and_is_true.py
+coverage run -a --source=gen_libs test/unit/gen_libs/binary_string.py
 coverage run -a --source=gen_libs test/unit/gen_libs/bytes_2_readable.py
 coverage run -a --source=gen_libs test/unit/gen_libs/chk_crt_dir.py
 coverage run -a --source=gen_libs test/unit/gen_libs/chk_crt_file.py