+++ CHANGELOG.md
@@ -9,11 +9,27 @@ Breaking Changes
 
 - Removed support for Python 2.7.
 
+### Added
+- gen_libs.binary_string: Returns lamdba as a function to check if data is binary.
+
 ### Changed
+- gen_libs.is_empty_file: Refactored the function.
+- gen_libs.file_search_cnt: Added "encoding" argument to open() command and set the "mode" to read to open() command.
+- gen_libs: Added "encoding" argument to open() command.
+- gen_libs: Converted strings to f-strings.
+- gen_libs.disk_usage: Changed from \_ntuple_diskusage to NTupleDiskUsage.
+- Renamed Tuple from \_ntuple_diskusage to ntuple_diskusage.
+- gen_libs: date_range, print_level, key_cleaner, prt_lvl: Cleaned up code.
+- gen_libs.merge_two_dicts: Replaced type() call with isinstance() call.
+- gen_libs.is_file_text: Moved lamdba statement to a binary_string function.
 - gen_libs.make_md5_hash: Removed check for Python 3.
 - gen_libs: is_base64, no_std_out: Remove Python 2.7 code.
 - gen_class.Dnf: Removed check for Python version for the class.
-- gen_class.daemonize: Remove Python 2.7 code.
+- gen_class.Daemonize: Remove Python 2.7 code.
+- gen_class.Daemonize2: Remove Python 2.7 code.
+- gen_class.LogFile: Remove Python 2.7 code and check for Python 2.7 version.
+- gen_class.KeyCaseInsensitiveDict.\_keylower: Replaced gen_libs.str_type with str.
+- gen_libs.merge_data_types: Replaced str_type() with str.
 
 ### Deprecated
 - gen_libs.str_type function.